:force option to overwrite output files if they exist

+++ src/file_tree.cr
@@ -12,16 +12,21 @@ module Teeplate
     def initialize(@__out_dir)
     end
 
-    def render
-      __try_to_write_files
+    def render(force = false)
+      __try_to_write_files force
     end
 
-    def __try_to_write(path, sliceable)
+    def __try_to_write(path, sliceable, force)
       path = File.join(@__out_dir, path)
       Dir.mkdir_p path.split("/")[0..-2].join("/")
+      return if File.exists?(path) && !__overwrite(path, sliceable, force)
       File.open(path, "w") do |f|
         f.write sliceable.to_slice
       end
+    end
+
+    def __overwrite(path, sliceable, force)
+      force
     end
   end
 end
